Remove bloat, reorganise/flatten directory structure and fix naming conventions, etc.
