+++ website/frontend/src/App.js
@@ -3,11 +3,11 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Home from './pages/Home';
 import Projects from './pages/Projects';
 import Navbar from './components/Navbar';
-import MyairopsProject from './pages/Projects/MyairopsProject';
-import EasyAProject from './pages/Projects/EasyAProject';
-import GoogleFormulaEProject from './pages/Projects/GoogleFormulaEProject';
-import HackawayV7Project from './pages/Projects/HackawayV7Project';
-import GoogleTagManager from './components/GoogleTagManager';
+import MyairopsProject from './pages/MyairopsProject';
+import EasyAProject from './pages/EasyAProject';
+import GoogleFormulaEProject from './pages/GoogleFormulaEProject';
+import HackawayV7Project from './pages/HackawayV7Project';
+import GoogleTagManager from './utilities/GoogleTagManager';
 
 function App() {
   return (
