+++ website/frontend/src/content/HackawayV7ProjectContent.js
