+++ website/frontend/src/pages/MyairopsProject.js
@@ -1,8 +1,7 @@
-import React from 'react';
 import { Helmet } from 'react-helmet';
 import ReactPlayer from 'react-player';
 import './MyairopsProject.css';
-import { MYAIROPS_ROLE, MYAIROPS_TECHNOLOGY_AND_SKILLS } from '../Content/Projects/MyairopsContent.js';
+import { MYAIROPS_ROLE, MYAIROPS_TECHNOLOGY_AND_SKILLS } from '../content/MyairopsProjectContent.js';
 
 function MyairopsProject() {
     return (
